My own branch

+++ Comment_microservice/src/main/java/karm/van/service/CommentService.java
@@ -3,10 +3,17 @@ package karm.van.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.annotation.PostConstruct;
-import karm.van.exception.CardNotFoundException;
-import karm.van.exception.CommentNotFoundException;
-import karm.van.exception.InvalidDataException;
-import karm.van.exception.SerializationException;
+import karm.van.config.AuthenticationMicroServiceProperties;
+import karm.van.dto.UserDtoRequest;
+import karm.van.exception.card.CardNotFoundException;
+import karm.van.exception.comment.CommentNotFoundException;
+import karm.van.exception.comment.CommentNotSavedException;
+import karm.van.exception.comment.CommentNotUnlinkException;
+import karm.van.exception.other.InvalidDataException;
+import karm.van.exception.other.SerializationException;
+import karm.van.exception.token.TokenNotExistException;
+import karm.van.exception.user.NotEnoughPermissionsException;
+import karm.van.exception.user.UsernameNotFoundException;
 import karm.van.model.CardModel;
 import karm.van.model.CommentModel;
 import karm.van.repo.CardRepo;
@@ -15,6 +22,10 @@ import karm.van.dto.CommentDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import redis.clients.jedis.JedisPooled;
@@ -29,9 +40,14 @@ public class CommentService {
     private final CommentRepo commentRepo;
     private final CardRepo cardRepo;
     private final ObjectMapper objectMapper;
+    private final AuthenticationMicroServiceProperties authProperties;
+    private final ApiService apiService;
 
     @Value("${redis.host}")
     private String redisHost;
+
+    @Value("${microservices.x-api-key}")
+    private String apiKey;
     private JedisPooled redis;
 
     @PostConstruct
@@ -39,8 +55,76 @@ public class CommentService {
         redis = new JedisPooled(redisHost,6379);
     }
 
+    private void checkToken(String token) throws TokenNotExistException {
+        if (!apiService.validateToken(token,
+                apiService.buildUrl(authProperties.getPrefix(),
+                        authProperties.getHost(),
+                        authProperties.getPort(),
+                        authProperties.getEndpoints().getValidateToken()
+                )
+        )){
+            throw new TokenNotExistException("Invalid token or expired");
+        }
+    }
+
+    public boolean checkNoneEqualsApiKey(String key) {
+        return !key.equals(apiKey);
+    }
+
+    private UserDtoRequest requestToGetUserByToken(String token) throws UsernameNotFoundException {
+        UserDtoRequest user = apiService.getUserByToken(apiService.buildUrl(
+                authProperties.getPrefix(),
+                authProperties.getHost(),
+                authProperties.getPort(),
+                authProperties.getEndpoints().getGetUserByToken()
+        ), token);
+
+        if (user == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+
+        return user;
+    }
+
+    private void requestToUnlinkCommentFromUser(String token, Long commentId, Long authorId) throws CommentNotUnlinkException {
+        HttpStatusCode httpStatusCode = apiService.requestToUnlinkCommentFromUser(apiService.buildUrl(
+                authProperties.getPrefix(),
+                authProperties.getHost(),
+                authProperties.getPort(),
+                authProperties.getEndpoints().getUnlinkCommentAndUser(),
+                commentId,
+                authorId
+        ),token,apiKey);
+
+        if (httpStatusCode != HttpStatus.OK){
+            throw new CommentNotUnlinkException("An error occurred while trying to delete the card");
+        }
+    }
+
+    private void requestToLinkCommentAndUser(Long commentId, String token) throws CommentNotSavedException {
+        String url = apiService.buildUrl(
+                authProperties.getPrefix(),
+                authProperties.getHost(),
+                authProperties.getPort(),
+                authProperties.getEndpoints().getLinkCommentAndUser(),
+                commentId);
+
+        try {
+            if (apiService.addCommentToUser(url,token,apiKey) != HttpStatus.OK){
+                log.error("The error occurred while the user was being assigned a card");
+                throw new CommentNotSavedException("An error occurred with saving");
+            }
+        }catch (NullPointerException | CommentNotSavedException e){
+            log.error("class: "+e.getClass()+" message: "+e.getMessage());
+            throw e;
+        }
+
+    }
+
     @Transactional
-    public void addComment(Long cardId, CommentDto commentDto) throws InvalidDataException, CardNotFoundException {
+    public void addComment(Long cardId, CommentDto commentDto, String authorization) throws InvalidDataException, CardNotFoundException, TokenNotExistException {
+        String token = authorization.substring(7);
+         checkToken(token);
         try {
             String commentText = commentDto.text();
 
@@ -51,48 +135,50 @@ public class CommentService {
             CardModel cardModel = cardRepo.getCardModelById(cardId)
                     .orElseThrow(()->new CardNotFoundException("Card with this id doesn't exist"));
 
+            UserDtoRequest user = requestToGetUserByToken(token);
+
             CommentModel commentModel = CommentModel.builder()
                     .text(commentDto.text())
                     .card(cardModel)
+                    .userId(user.id())
                     .createdAt(LocalDateTime.now())
                     .build();
 
             commentRepo.save(commentModel);
+
+            requestToLinkCommentAndUser(commentModel.getId(),token);
         } catch (InvalidDataException | CardNotFoundException e){
             throw e;
         } catch (Exception e){
-            log.debug("An unknown error occurred while deleting the card: "+e.getMessage()+" - "+e.getClass());
+            log.error("An unknown error occurred while deleting the card: "+e.getMessage()+" - "+e.getClass());
             throw new RuntimeException("Unexpected error occurred", e);
         }
     }
 
     private List<CommentModel> cacheComments(Long id,int limit,int page,String commentsKeyForCache) throws SerializationException {
-        List<CommentModel> comments = commentRepo.getCommentModelByCard_Id(id);// Идем в БД
+        Page<CommentModel> comments = commentRepo.getCommentModelByCard_Id(id, PageRequest.of(page,limit));// Идем в БД
 
         if (comments.isEmpty()){// Если комментариев вообще нет
             return List.of();//Возвращаем пустой список
         }
 
-        int start = page * limit; // Это индекс первого элемента (Если page = 1 (вторая страница) и limit = 10, то start = 1 * 10 = 10 — это будет 11-й элемент)
-        int end = Math.min(start + limit, comments.size());// Если на последней странице количество комментариев меньше, чем limit, то end будет равен количеству комментариев
-
-        List<CommentModel> pageComments = comments.subList(start, end);// Выбираем только нужное нам количество комментариев
-
-        for (CommentModel comment : pageComments) {// Проходимся по ним
-            try {
-                redis.rpush(commentsKeyForCache, objectMapper.writeValueAsString(comment));// Кешируем
-            } catch (JsonProcessingException e) {
-                throw new SerializationException("An error occurred during serialization");
-            }
-        }
+        comments.stream().parallel()
+                .forEach(comment->{
+                    try {
+                        redis.rpush(commentsKeyForCache, objectMapper.writeValueAsString(comment));// Кешируем
+                    } catch (JsonProcessingException e) {
+                        throw new RuntimeException(new SerializationException("An error occurred during serialization"));
+                    }
+                });
 
         redis.expire(commentsKeyForCache, 60); // Устанавливаем время жизни равное минуте
 
-        return pageComments;
+        return comments.getContent();
     }
 
-
-    public List<CommentModel> getComments(Long id,int limit,int page) throws CardNotFoundException, SerializationException {
+//TODO вместо id авторов получать их имена и данные
+    public List<CommentModel> getComments(Long id,int limit,int page, String authorization) throws CardNotFoundException, SerializationException, TokenNotExistException {
+        checkToken(authorization.substring(7));
         try {
             if (cardRepo.existsById(id)){// Проверяем существует ли такая карточка
                 String commentsKeyForCache = "comments:card:"+id; // Ключ в редисе от списка комментариев
@@ -124,7 +210,9 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteAllCommentsByCard(Long cardId){
+    public void deleteAllCommentsByCard(Long cardId, String authorization) throws TokenNotExistException {
+        String token = authorization.substring(7);
+        checkToken(token);
         try {
             List<CommentModel> comments = commentRepo.getCommentModelByCard_Id(cardId);
             String commentsKeyForCache = "comments:card:%d".formatted(cardId);
@@ -134,6 +222,13 @@ public class CommentService {
             }
 
             if (!comments.isEmpty()){
+                comments.parallelStream().forEach(comment->{
+                    try {
+                        requestToUnlinkCommentFromUser(token,comment.getId(),comment.getUserId());
+                    } catch (CommentNotUnlinkException e){
+                        throw new RuntimeException(e.getMessage());
+                    }
+                });
                 commentRepo.deleteAllByCard_Id(cardId);
             }
         } catch (Exception e){
@@ -143,8 +238,19 @@ public class CommentService {
 
     }
 
+    private void checkPermissions(CommentModel commentModel, String token) throws UsernameNotFoundException, NotEnoughPermissionsException {
+        Long commentAuthorId = commentModel.getUserId();
+        UserDtoRequest user = requestToGetUserByToken(token);
+
+        if (!commentAuthorId.equals(user.id()) && user.role().stream().noneMatch(role->role.equals("ADMIN"))){
+            throw new NotEnoughPermissionsException("You don't have permission to do this");
+        }
+    }
+
     @Transactional
-    public void patchComment(Long commentId, CommentDto commentDto) throws InvalidDataException, CommentNotFoundException {
+    public void patchComment(Long commentId, CommentDto commentDto, String authorization) throws InvalidDataException, CommentNotFoundException, TokenNotExistException {
+        String token = authorization.substring(7);
+        checkToken(token);
         try {
             String commentText = commentDto.text();
 
@@ -154,6 +260,8 @@ public class CommentService {
 
             CommentModel commentModel = commentRepo.getCommentModelById(commentId)
                     .orElseThrow(()->new CommentNotFoundException("Card with this id doesn't exist"));
+
+            checkPermissions(commentModel,token);
 
             commentModel.setText(commentText);
 
@@ -167,7 +275,13 @@ public class CommentService {
     }
 
     @Transactional
-    public void deleteOneComment(Long commentId) {
+    public void deleteOneComment(Long commentId, String authorization) throws TokenNotExistException, CommentNotFoundException, UsernameNotFoundException, NotEnoughPermissionsException, CommentNotUnlinkException {
+        String token = authorization.substring(7);
+        checkToken(token);
+        CommentModel commentModel = commentRepo.getCommentModelById(commentId)
+                .orElseThrow(()->new CommentNotFoundException("Card with this id doesn't exist"));
+        checkPermissions(commentModel,token);
+        requestToUnlinkCommentFromUser(token,commentId,commentModel.getUserId());
         commentRepo.deleteById(commentId);
     }
 }
